Fix prettier config (revert to prettier v2), fix scroll disabling for mobile menu, and fix contributor loading
Contributor loading worked, but Svelte complained that it was happening during serverside
rendering (so the site might have to be redeployed to get contributors to update). The prettier
config just straight up didn't work so I had no automatic formatter, but now it works, and
running it formatted a lot of stuff. The svelte:body approach to disabling scrolling while
the mobile nav menu is open didn't work (it seems like svelte:body doesn't really allow
styling in that dynamic way)